+++ site/src/main/scala/org/opentorah/site/SiteService.scala
@@ -1,52 +1,36 @@
 package org.opentorah.site
 
-import cats.data.OptionT
-import fs2.Stream
 import net.logstash.logback.argument.{StructuredArgument, StructuredArguments}
-import org.http4s.dsl.Http4sDsl
-import org.http4s.headers.{`Content-Length`, `Content-Type`}
-import org.http4s.blaze.client.BlazeClientBuilder
-import org.http4s.blaze.server.BlazeServerBuilder
-import org.http4s.{Charset, Headers, HttpRoutes, MediaType, Response, Request, StaticFile, Uri}
-import org.slf4j.{Logger, LoggerFactory}
-import org.opentorah.util.{Effects, Files}
+import io.netty.handler.codec.http.HttpHeaderNames
+import org.slf4j.Logger
+import org.opentorah.util.{Effects, Files, Logging, Zhttp}
 import org.opentorah.xml.{Element, Parser}
-import org.typelevel.ci.CIString
+import zhttp.http._
 import zio.duration.Duration
-import zio.interop.catz._
-import zio.{RIO, Task, URIO, ZEnv}
-import java.io.File
-import java.net.URL
-import scala.concurrent.{ExecutionContext, ExecutionContextExecutor}
-import scala.jdk.CollectionConverters.SeqHasAsJava
+import zio.stream.ZStream
+import zio.{Chunk, Task, ZEnv, ZIO}
+import zio.blocking.Blocking
 
-// see https://zio.dev/docs/howto/interop/with-cats-effect
-abstract class SiteService[S <: Site[S]] extends Element[S]("site") with zio.interop.catz.CatsApp {
+import java.io.File
+import java.net.{URI, URL}
+
+
+abstract class SiteService[S <: Site[S]] extends Element[S]("site") with zio.App {
 
   // This is supposed to be set when running in Cloud Run
   private val serviceName: Option[String] = Option(System.getenv("K_SERVICE"))
 
-  LoggerFactory.getILoggerFactory match {
-    case loggerContext: ch.qos.logback.classic.LoggerContext =>
-      SiteService.configureLogback(loggerContext, inCloudRun = serviceName.isDefined)
-
-    case _ =>
-  }
+  Logging.configureLogBack(useLogStash = serviceName.isDefined)
 
   private type ServiceEnvironment = ZEnv
 
-  private type ServiceTask[+A] = RIO[ServiceEnvironment, A]
-
   private final def readSite(url: URL): Task[S] = readSiteFile(Files.fileInDirectory(url, "site.xml"))
 
-  private final def readSiteFile(url: URL): Task[S] = Parser.toTask(
-    Effects.effect(logger.info(s"Reading site from $url")) *>
-    parse(url)
-  )
-
-  // TODO remove?
-//  private final def doReadSiteFile(file: File): S = Parser.run(readSiteFile(file))
-//  private final def readSiteFile(file: File): Parser[S] = parse(Files.file2url(file))
+  private final def readSiteFile(url: URL): Task[S] = Parser.toTask(for {
+    _ <- Effects.effect(logger.info(s"Reading site from $url"))
+    result <- parse(url)
+    _ <- Effects.effect(logger.info(s"Reading site from $url - done"))
+  } yield result)
 
   // TODO HTTP GET from http://metadata.google.internal/computeMetadata/v1/project/project-id
   // with `Metadata-Flavor: Google` header;
@@ -56,7 +40,7 @@ abstract class SiteService[S <: Site[S]] extends Element[S]("site") with zio.int
   // TODO does this belong in the base class?
   protected def bucketName: String
 
-  final override def run(args: List[String]): URIO[ServiceEnvironment, zio.ExitCode] = {
+  final override def run(args: List[String]): zio.URIO[ServiceEnvironment, zio.ExitCode] = {
     if (args.isEmpty) serve(urlString = None) else args.head match {
       case "serveRemoteSite"         => serve(urlString = None)
       case "serveSite"               => serve(urlString = args.lift(1))
@@ -69,7 +53,7 @@ abstract class SiteService[S <: Site[S]] extends Element[S]("site") with zio.int
 
   private def build(urlString: String, withPrettyPrint: Boolean): Task[Unit] = {
     val url: URL = Files.file2url(new File(urlString))
-    readSite(url) >>= (_.build(withPrettyPrint))
+    readSite(url).flatMap(_.build(withPrettyPrint))
   }
 
   private def upload(directoryPath: String, serviceAccountKey: String, dryRun: Boolean): Task[Unit] =
@@ -81,8 +65,8 @@ abstract class SiteService[S <: Site[S]] extends Element[S]("site") with zio.int
       dryRun = dryRun
     ).sync())
 
-  private def serve(urlString: Option[String]): URIO[ServiceEnvironment, zio.ExitCode] = {
-    def getParameter(name: String, defaultValue: String): String = Option(System.getenv(name)).fold {
+  private def serve(urlString: Option[String]): zio.URIO[ServiceEnvironment, zio.ExitCode] = {
+    def getParameter(name: String, defaultValue: String): String = scala.util.Properties.envOrNone(name).fold {
       info(s"No value for '$name' in the environment; using default: '$defaultValue'")
       defaultValue
     }{ value =>
@@ -100,162 +84,106 @@ abstract class SiteService[S <: Site[S]] extends Element[S]("site") with zio.int
 
     warning(s"serviceName=$serviceName")
 
-    stream(port, siteUrl)
-      .compile
-      .drain
-      .mapError(err => zio.console.putStrLn(s"Execution failed with: $err"))
-      .exitCode
+    Zhttp.start(port, routes(siteUrl))
   }
 
-  private def stream(port: Int, siteUrl: String): Stream[ServiceTask, Nothing] = {
-    import org.http4s.implicits._
-
-    val executionContext: ExecutionContextExecutor = ExecutionContext.global
-    val httpApp = routes(siteUrl).orNotFound
-
-    val result = for {
-      _ <- BlazeClientBuilder[ServiceTask](executionContext).stream
-      exitCode <- BlazeServerBuilder[ServiceTask](executionContext)
-        // To be accessible when running in a docker container the server
-        // must bind to all IPs, not just 127.0.0.1:
-        .bindHttp(port, "0.0.0.0")
-        .withWebSockets(false)
-        .withHttpApp(httpApp)
-        .serve : Stream[ServiceTask, cats.effect.ExitCode]
-    } yield exitCode
-
-    // TODO Scala 3.0.2 requires this cast (there is some ascription workaround...), but Scala 2 does not?!
-    (result.asInstanceOf[Stream[ServiceTask, cats.effect.ExitCode]])
-      .drain
-  }
-
-  private def routes(siteUrl: String): HttpRoutes[ServiceTask] = {
-    val dsl = new Http4sDsl[ServiceTask] {}
-    import dsl._
-
-    val siteUri: Uri = Uri.unsafeFromString(siteUrl)
+  private def routes(siteUrl: String): RHttpApp[ServiceEnvironment] = {
+    val siteUri: URI = URI.create(siteUrl)
 
     // TODO move near the SiteReader
     var cachedSite: Option[S] = None
-    def getSite: Task[S] = cachedSite.map(Task.succeed(_)).getOrElse(readSite(SiteService.toUrl(siteUri)).map { result =>
+    def getSite: Task[S] = cachedSite.map(Task.succeed(_)).getOrElse(readSite(siteUri.toURL).map { result =>
       cachedSite = Some(result)
       result
     })
 
     Effects.unsafeRun(getSite)
 
-    def get(request: Request[ServiceTask]): ServiceTask[Response[ServiceTask]] = {
-      // TODO if I add a parameter list (implicit F: Sync[ServiceTask]) and use F.pure and F.suspend, I get
-      //   ambiguous implicit values:
-      //   both method taskConcurrentInstance in trait CatsEffectInstances1 of type
-      //     [R]cats.effect.Concurrent[[β$18$]zio.ZIO[R,Throwable,β$18$]]
-      //   and value F of type cats.effect.Sync[org.opentorah.collector.Service.ServiceTask]
-      //   match expected type cats.effect.Sync[org.opentorah.collector.Service.ServiceTask]
-      //val F: Sync[ServiceTask] = taskConcurrentInstance
+    def reset(request: Request): Task[UResponse] = for {
+      _ <- Effects.effectTotal({
+        info(request, "RST")
+        cachedSite = None
+      })
+      _ <- getSite
+    } yield Response.http(
+      headers = List(Header.contentTypeTextPlain),
+      content = Zhttp.textData("Site reset!")
+    )
 
-      val path: Seq[String] = Files.splitAndDecodeUrl(request.uri.path.toString)
+    def get(request: Request): ResponseM[ServiceEnvironment, Throwable] = {
+      val path: Seq[String] = Files.splitAndDecodeUrl(request.url.path.asString)
       val urlStr: String = Files.mkUrl(path)
 
-      OptionT[ServiceTask, Response[ServiceTask]](for {
+      val result = for {
         site <- getSite
         siteResponse <- site.resolveContent(path)
-        result = siteResponse.map { siteResponse =>
-          val bytes: Array[Byte] = siteResponse.content.getBytes
-          Response[ServiceTask](
-            headers = Headers(
-              `Content-Type`(MediaType.unsafeParse(siteResponse.mimeType), Charset.`UTF-8`),
-              `Content-Length`.unsafeFromLong(bytes.length)
+        result <- siteResponse.map(siteResponse => ZIO.succeed {
+          val bytes: Array[Byte] = Zhttp.textBytes(siteResponse.content)
+
+          Response.http(
+            headers = List(
+              // TODO: `Content-Type`(MediaType.unsafeParse(siteResponse.mimeType), Charset.`UTF-8`)
+              Header(HttpHeaderNames.CONTENT_TYPE, siteResponse.mimeType),
+              Header.contentLength(bytes.length.toLong)
               // TODO more headers!
             ),
-            //  // TODO Scala 3.0.2 requires this cast (but there is some)
-            // TODO Scala 3.0.2 requires this cast, but Scala 2 does not?!
-            body = Stream.emits(bytes).asInstanceOf[Stream[ServiceTask, Byte]]
+            content = HttpData.fromStream(ZStream.fromChunk(Chunk.fromArray(bytes)))
           )
+        }).getOrElse {
+          val path: String = request.url.path.asString
+          for {
+            static <- Zhttp.staticFile(URI.create(s"$siteUrl$path").toURL, Some(request))
+          } yield static.getOrElse(Zhttp.notFound(path))
         }
-      } yield result)
-        .orElse {
-          val path: Uri.Path = request.uri.path
-          val implied: Uri.Path = if (path.endsWithSlash) path/Uri.Path.Segment("index.html") else path
-          val uri: Uri = siteUri.copy(path = siteUri.path.addSegments(implied.segments))
-          StaticFile.fromURL[ServiceTask](SiteService.toUrl(uri), Some(request))
+      } yield result
+
+      result.timed.mapEffect { case (duration: Duration, response: Response[Blocking, Throwable]) =>
+        val durationStr: String = SiteService.formatDuration(duration)
+
+        val isOk: Boolean = response match {
+          case response: Response.HttpResponse[Blocking, Throwable] => response.status == Status.OK
+          case _ => false
         }
-        .getOrElseF(NotFound(s"Not found: $urlStr"))
-        .timed.mapEffect { case (duration: Duration, response: Response[ServiceTask]) =>
-          val durationStr: String = SiteService.formatDuration(duration)
 
-          if (response.status.isInstanceOf[NotFound.type])
-            warning(request, s"NOT $durationStr $urlStr")
-          else
-            info   (request, s"GOT $durationStr $urlStr")
+        if (isOk)
+          info   (request, s"GOT $durationStr $urlStr")
+        else
+          warning(request, s"NOT $durationStr $urlStr")
 
-          response
+        response
       }
     }
 
-    HttpRoutes.strict[ServiceTask] {
-      case request@GET -> Root / "reset-cached-site" => for {
-        _ <- Effects.effectTotal({
-          info(request, "RST")
-          cachedSite = None
-        })
-        _ <- getSite
-        result <- Ok("Site reset!")
-      } yield result
-
-      case request@GET -> _ => get(request)
+    Http.collectM[Request] {
+      case request@Method.GET -> Root / "reset-cached-site" => reset(request)
+      case request => get(request)
     }
   }
 
-  private def info   (request: Option[Request[ServiceTask]], message: String): Unit = log(request, message, "INFO"   )
-  private def info   (request: Request[ServiceTask], message: String): Unit = log(Some(request), message, "INFO"   )
-  private def info   (                               message: String): Unit = log(None         , message, "INFO"   )
-  //  private def notice (request: Request[ServiceTask], message: String): Unit = log(Some(request), message, "NOTICE" )
-  //  private def notice (                               message: String): Unit = log(None         , message, "NOTICE" )
-  private def warning(request: Request[ServiceTask], message: String): Unit = log(Some(request), message, "WARNING")
-  private def warning(                               message: String): Unit = log(None         , message, "WARNING")
+  // TODO move into Logging?
+
+  private def info   (request: Option[Request], message: String): Unit = log(     request , message, "INFO"   )
+  private def info   (request:        Request , message: String): Unit = log(Some(request), message, "INFO"   )
+  private def info   (                          message: String): Unit = log(None         , message, "INFO"   )
+  private def notice (request:        Request , message: String): Unit = log(Some(request), message, "NOTICE" )
+  private def notice (                          message: String): Unit = log(None         , message, "NOTICE" )
+  private def warning(request:        Request , message: String): Unit = log(Some(request), message, "WARNING")
+  private def warning(                          message: String): Unit = log(None         , message, "WARNING")
 
   private final def logger: Logger = Site.logger
 
-  private def log(request: Option[Request[ServiceTask]], message: String, severity: String): Unit = {
-    val trace: Option[String] = request.flatMap(_
-      .headers.get(CIString("X-Cloud-Trace-Context"))
-      .map(_.toString.split("/")(0))
-    )
+  private def log(request: Option[Request], message: String, severity: String): Unit = {
+    val trace: Option[String] = request.flatMap(_.getHeaderValue("X-Cloud-Trace-Context")).map(_.split("/")(0))
 
     val arguments: Seq[StructuredArgument] =
       Seq(StructuredArguments.keyValue("severity", severity)) ++
-        trace.toSeq.map(trace => StructuredArguments.keyValue("logging.googleapis.com/trace", s"projects/$projectId/traces/$trace"))
+      trace.toSeq.map(trace => StructuredArguments.keyValue("logging.googleapis.com/trace", s"projects/$projectId/traces/$trace"))
 
     logger.info(message, arguments: _*)
   }
 }
 
 object SiteService {
-
-  private def configureLogback(loggerContext: ch.qos.logback.classic.LoggerContext, inCloudRun: Boolean): Unit = {
-    val rootLogger: ch.qos.logback.classic.Logger = loggerContext.getLogger(Logger.ROOT_LOGGER_NAME)
-
-    if (inCloudRun) {
-      val statusManager = loggerContext.getStatusManager
-      if (statusManager != null) statusManager.add(new ch.qos.logback.core.status.InfoStatus("Configuring logger", loggerContext))
-
-      val encoder = new net.logstash.logback.encoder.LogstashEncoder
-      // Ignore default logging fields
-      encoder.setExcludeMdcKeyNames(List("timestamp", "version", "logger", "thread", "level", "levelValue").asJava)
-
-      val consoleAppender = new ch.qos.logback.core.ConsoleAppender[ch.qos.logback.classic.spi.ILoggingEvent]
-      consoleAppender.setName("jsonConsoleAppender")
-      consoleAppender.setContext(loggerContext)
-      consoleAppender.setEncoder(encoder)
-
-      rootLogger.detachAndStopAllAppenders()
-      rootLogger.addAppender(consoleAppender)
-    }
-
-    rootLogger.setLevel(ch.qos.logback.classic.Level.INFO)
-  }
-
-  private def toUrl(uri: Uri): URL = new URL(uri.toString)
 
   private def formatDuration(duration: Duration): String = {
     val millis: Long = duration.toMillis