+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+- switched from http4s to zhttp;
 
 ## [0.2.6] - 2021-08-17
 - consolidated 'util', 'xml' and 'metadata' modules into 'base';